Replace toggle + add buttons with dedicated whitelist/graylist buttons.
This improves the UI in a couple ways: now there is only one click to add either
to the whitelist/graylist and there is no non-obvious toggle button to switch states.
Furthermore, the "primary" (whitelist) button is more visually emphasized and is the
default action when you press Enter.

This also removes the redundant header on the input and replaces it with an inline
placeholder to lead by example.