Make the regExp property on expressions computed instead of stored.
Previously, the stored value was just a cached version of the computed value. This
is prone to correctness issues where one is accidentally not updated. Since the
regexp is only for display purposes and is not actually editable, it should always
be computed fresh to make sure it's not accidentally out of date.

This will also make it safer to dynamically update the regexp preview when editing.