Added Documenation and FAQ Links (GitHub).
NOTE:  Locales will need to be updated as additional texts are included.  Actually there may be additional locale changes down the line that will need to be adjusted/re-translated again while I try to remove a minor format duplication.

Also added in an update to the cookie counter that should not count the localstoragecleanup cookie created by CAD.

Last but not least minor dependency updates.

Signed-off-by: Kenneth Tran <kennethtran93@users.noreply.github.com>