Let Firefox on Android manage all storage types.
Firefox on Android now supports more than just cookies so the extension
should allow users to manage those as well.

Adds a method to generate a map of the current supported storage types
so that they can be easily queried. This isolates all the version
checking to one place as well.